+++ pkg/webui/console/containers/gateway-onboarding-form/gateway-provisioning-form/gateway-registration-form-section/index.js
@@ -16,8 +16,6 @@ import React from 'react'
 import { useFormikContext } from 'formik'
 import { defineMessages } from 'react-intl'
 
-import frequencyPlans from '@console/constants/frequency-plans'
-
 import Form from '@ttn-lw/components/form'
 import Input from '@ttn-lw/components/input'
 import Checkbox from '@ttn-lw/components/checkbox'
@@ -28,7 +26,7 @@ import SubmitButton from '@ttn-lw/components/submit-button'
 
 import Message from '@ttn-lw/lib/components/message'
 
-import { GsFrequencyPlansSelect as FrequencyPlansSelect } from '@console/containers/freq-plans-select'
+import GsFrequencyPlansSelect from '@console/containers/freq-plans-select/gs-frequency-plan-select'
 
 import { selectGsConfig } from '@ttn-lw/lib/selectors/env'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
@@ -67,7 +65,7 @@ const initialValues = {
   },
   status_public: true,
   location_public: true,
-  frequency_plan_id: '',
+  frequency_plan_ids: [''],
   name: '',
   require_authenticated_connection: false,
   gateway_server_address: gsEnabled ? new URL(gsBaseURL).hostname : '',
@@ -95,19 +93,7 @@ const GatewayRegistrationFormSections = () => {
         component={Input}
         tooltipId={tooltipIds.GATEWAY_NAME}
       />
-      {gsEnabled && (
-        <FrequencyPlansSelect
-          name="frequency_plan_id"
-          menuPlacement="top"
-          tooltipId={tooltipIds.FREQUENCY_PLAN}
-          warning={
-            values.frequency_plan_id === frequencyPlans.EMPTY_FREQ_PLAN
-              ? sharedMessages.frequencyPlanWarning
-              : undefined
-          }
-          required
-        />
-      )}
+      {gsEnabled && <GsFrequencyPlansSelect />}
       <Form.Field
         name="require_authenticated_connection"
         component={Checkbox}