+++ pkg/webui/console/views/gateway-general-settings/lorawan-settings-form/validation-schema.js
@@ -18,7 +18,7 @@ import sharedMessages from '@ttn-lw/lib/shared-messages'
 import { delay as delayRegexp } from '@console/lib/regexp'
 
 const validationSchema = Yup.object().shape({
-  frequency_plan_id: Yup.string(),
+  frequency_plan_ids: Yup.array(),
   schedule_downlink_late: Yup.boolean().default(false),
   schedule_anytime_delay: Yup.string().matches(
     delayRegexp,